Fixing issue PR5
