+++ src/components/core/CSModal.tsx
@@ -4,7 +4,7 @@ import Icon from 'react-native-vector-icons/Feather';
 import RBSheet from 'react-native-raw-bottom-sheet';
 import {SPACING} from '../../constants/spacing';
 import {COLORS} from '../../constants/color';
-import CSLayout from './CSLayout';
+import {CSLayout} from './CSLayout';
 
 interface CSModalProps {
   refRBSheet: any;
@@ -22,29 +22,8 @@ const CSModal = (props: CSModalProps) => {
       closeOnDragDown={false}
       animationType="slide"
       customStyles={{
-        container: {
-          justifyContent: 'flex-start',
-          alignItems: 'center',
-          height: height,
-          borderRadius: 15,
-          width: SPACING.screenWidth - SPACING.px * 2,
-          position: 'absolute',
-          shadowColor: COLORS.black,
-          backgroundColor: COLORS.bgDark,
-          shadowOffset: {
-            width: 0,
-            height: 3,
-          },
-          shadowOpacity: 0.27,
-          shadowRadius: 4.65,
-
-          elevation: 6,
-          ...style,
-        },
-        wrapper: {
-          justifyContent: 'center',
-          alignItems: 'center',
-        },
+        container: [styles.containerRBS, {height: height, ...style}],
+        wrapper: styles.wrapper,
       }}>
       {closeBtn && (
         <TouchableOpacity
@@ -74,6 +53,27 @@ const styles = StyleSheet.create({
     paddingTop: 40,
     gap: 10,
   },
+  containerRBS: {
+    justifyContent: 'flex-start',
+    alignItems: 'center',
+    borderRadius: 15,
+    width: SPACING.screenWidth - SPACING.px * 2,
+    position: 'absolute',
+    shadowColor: COLORS.black,
+    backgroundColor: COLORS.bgDark,
+    shadowOffset: {
+      width: 0,
+      height: 3,
+    },
+    shadowOpacity: 0.27,
+    shadowRadius: 4.65,
+
+    elevation: 6,
+  },
+  wrapper: {
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
 });
 
-export default CSModal;
+export {CSModal};
