+++ src/components/course/GroupCourseRender.tsx
@@ -2,7 +2,7 @@ import {StyleSheet, View} from 'react-native';
 import React from 'react';
 
 import {CourseItem} from '../../types';
-import {CircleProgress, Line, MainElement} from './GroupCourseRenderMaterial';
+import {CircleProgress, Line, MainElement} from './renderMaterial';
 
 interface GroupCourseRenderProps {
   groupItem: CourseItem[];
@@ -13,50 +13,43 @@ const GroupCourseRender = ({
   groupItem,
   filtedArray,
 }: GroupCourseRenderProps) => {
+  const [item1, ...item23] = groupItem;
   return (
     <>
-      {groupItem.map((item, index) => {
-        if (index === 0) {
+      <View style={styles.first} key={item1.id}>
+        <MainElement
+          cover={item1.cover}
+          learnedLesson={item1.learnedLesson}
+          name={item1.name}
+          totalLesson={item1.totalLesson}
+        />
+        {item1.id !== filtedArray[0].id && (
+          <Line variant="line12" isActive={item1.learnedLesson > 0} />
+        )}
+        <CircleProgress filteredArray={filtedArray} courseItem={item1} />
+      </View>
+
+      <View style={styles.secondAndThird}>
+        {item23.map((item, index) => {
           return (
-            <View style={styles.first} key={item.id}>
+            <View key={index} style={styles.item23}>
               <MainElement
                 cover={item.cover}
                 learnedLesson={item.learnedLesson}
                 name={item.name}
                 totalLesson={item.totalLesson}
               />
-              {item.id !== filtedArray[0].id && (
-                <Line variant="line12" isActive={item.learnedLesson > 0} />
-              )}
-              <CircleProgress filteredArray={filtedArray} courseItem={item} />
+              <Line
+                variant={index === 1 ? 'line3' : 'line12'}
+                isActive={item.learnedLesson > 0}
+              />
+              <CircleProgress
+                filteredArray={filtedArray}
+                courseItem={item}
+                isThirdCircle={index === 1}
+              />
             </View>
           );
-        }
-      })}
-
-      <View style={styles.secondAndThird}>
-        {groupItem.map((item, index) => {
-          if (index !== 0) {
-            return (
-              <View key={index} style={styles.item23}>
-                <MainElement
-                  cover={item.cover}
-                  learnedLesson={item.learnedLesson}
-                  name={item.name}
-                  totalLesson={item.totalLesson}
-                />
-                <Line
-                  variant={index === 2 ? 'line3' : 'line12'}
-                  isActive={item.learnedLesson > 0}
-                />
-                <CircleProgress
-                  filteredArray={filtedArray}
-                  courseItem={item}
-                  isThirdCircle={index === 2}
-                />
-              </View>
-            );
-          }
         })}
       </View>
     </>
