+++ src/components/course/renderMaterial/index.ts
@@ -1,0 +1,3 @@
+export * from './CircleProgress';
+export * from './Line';
+export * from './MainElement';
