+++ src/screens/auth/Login.tsx
@@ -3,14 +3,11 @@ import React, {useState} from 'react';
 import {Link, useNavigation} from '@react-navigation/native';
 import Parse from 'parse/react-native';
 
-import CSText from '../../components/core/CSText';
-import CSInput from '../../components/core/CSInput';
 import {CSButton} from '../../components/core/CSButton';
 import {SPACING} from '../../constants/spacing';
-import CSLoading from '../../components/core/CSLoading';
 import {AppDispatch} from '../../store/store';
 import {USER_ACTION} from '../../store/actions';
-import CSLayout from '../../components/core/CSLayout';
+import {CSInput, CSLayout, CSLoading, CSText} from '../../components/core';
 
 const Login = () => {
   const navigation = useNavigation<any>();
