+++ src/screens/game/GameRollDice.tsx
@@ -1,7 +1,5 @@
 import React from 'react';
-
-import CSText from '../../components/core/CSText';
-import CSLayout from '../../components/core/CSLayout';
+import {CSLayout, CSText} from '../../components/core';
 
 const GameRollDice = () => {
   return (
