+++ src/screens/library/Library.tsx
@@ -2,8 +2,7 @@ import {useNavigation} from '@react-navigation/native';
 import React, {useEffect} from 'react';
 
 import HeaderScreen from '../../components/HeaderScreen';
-import CSText from '../../components/core/CSText';
-import CSLayout from '../../components/core/CSLayout';
+import {CSLayout, CSText} from '../../components/core';
 
 const Library = () => {
   const navigation = useNavigation();
